fix: bump contract and permissioned integration (#87)
Accompanying with https://github.com/babylonlabs-io/babylon/pull/291

This PR

- bumps contract against the latest Babylon base branch
- fixes all failed tests due to the bump
- implements the consumer side logic of permissioned integration, mostly
removing auto consumer registration

+++ contracts/babylon/src/state/config.rs
@@ -9,8 +9,8 @@ pub(crate) const CONFIG: Item<Config> = Item::new("config");
 pub struct Config {
     pub network: babylon_bitcoin::chain_params::Network,
     pub babylon_tag: Vec<u8>,
-    pub btc_confirmation_depth: u64,
-    pub checkpoint_finalization_timeout: u64,
+    pub btc_confirmation_depth: u32,
+    pub checkpoint_finalization_timeout: u32,
     /// notify_cosmos_zone indicates whether to send Cosmos zone messages notifying BTC-finalised headers.
     /// NOTE: if set to true, then the Cosmos zone needs to integrate the corresponding message
     /// handler as well
